Move Button and Label to a Separate File
Moved all the functions for UIButtons and labels to a separate file called uielements. This allows them to be used multiple times in the project if a different screen gets implemented and needs them.
